[installer]: refactor yq to allow it run scripts against strings
